Refactor engine library
