+++ Dependencies/RocketEngine/src/GUI/text_builder.cpp
@@ -1,9 +1,9 @@
-#include "text_builder.h"
+#include "GUI/text_builder.h"
 
 #include <cstdlib>
 
-#include "bmp/loadbmp.h"
-#include "memory/psp_malloc.h"
+#include "loadbmp.h"
+#include "psp_malloc.h"
 
 Animation font = Animation(9, 9, 0, "assets/misc/font.bmp");
 bool initialized = false;
