+++ Dependencies/RocketEngine/src/loadbmp.cpp
@@ -4,8 +4,8 @@
 #include <stdlib.h>
 #include <math.h>
 
-#include "memory/psp_malloc.h"
-#include "logger/logger.h"
+#include "psp_malloc.h"
+#include "logger.h"
 
 int bmp_mem=0;
