+++ Dependencies/RocketEngine/src/physics/collisions.cpp
@@ -1,7 +1,7 @@
 #include "globals.h"
 
 #include "object/object_list.h"
-#include "logger/logger.h"
+#include "logger.h"
 #include "physics/collisions.h"
 
 int object_collision(Object * obj, ObjectList &collisions, std::set<Object::ObjectTypes> ignore) {
