+++ src/game/object_handler/animation_handler.cpp
@@ -1,6 +1,6 @@
 #include "animation_handler.h"
-#include "graphics/gfx.hpp"
-#include "bmp/loadbmp.h"
+#include "gfx.hpp"
+#include "loadbmp.h"
 
 AnimationHandler::AnimationHandler(int MAX_OBJECTS, int _velocity,
     int _time_between_spawns, int * camera_x, unsigned char * terrain,
