+++ src/game/object_handler/explosion_handler.h
@@ -1,4 +1,4 @@
-#include "animation_handler.h"
+#include "object_handler/animation_handler.h"
 class ExplosionHandler : public AnimationHandler
 {
 public:
